requirements.txt
# Dependencias principales para Tráfico Candy Gemini (Telegram)
python-telegram-bot>=20.8
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0

# Dependencias para automatización web (plataformas)
selenium>=4.15.0
webdriver-manager>=4.0.0
pyautogui>=0.9.54

# Dependencias para tipado y utilidades
typing-extensions>=4.8.0

# Dependencias opcionales para funcionalidad extendida
pillow>=10.1.0
python-dateutil>=2.8.2
pytz>=2023.3

# Supabase para base de datos
supabase>=2.0.0
